Refactoring the code. Splitted the MAS class to MASActions
